Define Transaction interface in storage package

See merge request https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7242

Merged-by: James Liu <jliu@gitlab.com>
Approved-by: James Liu <jliu@gitlab.com>

+++ internal/gitaly/service/repository/remove.go
@@ -5,7 +5,6 @@ import (
 
 	"gitlab.com/gitlab-org/gitaly/v16/internal/gitaly/repoutil"
 	"gitlab.com/gitlab-org/gitaly/v16/internal/gitaly/storage"
-	"gitlab.com/gitlab-org/gitaly/v16/internal/gitaly/storage/storagectx"
 	"gitlab.com/gitlab-org/gitaly/v16/internal/structerr"
 	"gitlab.com/gitlab-org/gitaly/v16/proto/go/gitalypb"
 )
@@ -20,9 +19,9 @@ func (s *server) RemoveRepository(ctx context.Context, in *gitalypb.RemoveReposi
 		return nil, err
 	}
 
-	storagectx.RunWithTransaction(ctx, func(tx storagectx.Transaction) {
+	if tx := storage.ExtractTransaction(ctx); tx != nil {
 		tx.DeleteRepository()
-	})
+	}
 
 	return &gitalypb.RemoveRepositoryResponse{}, nil
 }
